[Style] StartView, StartViewController 분리
